# apps/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gun Del Sol API",
    "description": "High-performance async API for Solana token analysis (Modular)",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/tokens/{token_id}": {
      "get": {
        "tags": [
          "Tokens"
        ],
        "summary": "Get Token By Id",
        "description": "Get token details with wallets and axiom export",
        "operationId": "get_token_by_id_api_tokens__token_id__get",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tokens"
        ],
        "summary": "Soft Delete Token",
        "description": "Soft delete a token (move to trash)",
        "operationId": "soft_delete_token_api_tokens__token_id__delete",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tokens/{token_id}/history": {
      "get": {
        "tags": [
          "Tokens"
        ],
        "summary": "Get Token Analysis History",
        "description": "Get analysis history for a specific token",
        "operationId": "get_token_analysis_history_api_tokens__token_id__history_get",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisHistory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tokens/{token_id}/restore": {
      "post": {
        "tags": [
          "Tokens"
        ],
        "summary": "Restore Token",
        "description": "Restore a soft-deleted token",
        "operationId": "restore_token_api_tokens__token_id__restore_post",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tokens/{token_id}/permanent": {
      "delete": {
        "tags": [
          "Tokens"
        ],
        "summary": "Permanent Delete Token",
        "description": "Permanently delete a token and all associated data",
        "operationId": "permanent_delete_token_api_tokens__token_id__permanent_delete",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tokens/refresh-market-caps": {
      "post": {
        "tags": [
          "Tokens"
        ],
        "summary": "Refresh Market Caps",
        "description": "Refresh current market cap for multiple tokens",
        "operationId": "refresh_market_caps_api_tokens_refresh_market_caps_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshMarketCapsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshMarketCapsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisHistory": {
        "properties": {
          "token_id": {
            "type": "integer",
            "title": "Token Id"
          },
          "total_runs": {
            "type": "integer",
            "title": "Total Runs"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/AnalysisRun"
            },
            "type": "array",
            "title": "Runs"
          }
        },
        "type": "object",
        "required": [
          "token_id",
          "total_runs",
          "runs"
        ],
        "title": "AnalysisHistory"
      },
...
      "AnalysisRun": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "analysis_timestamp": {
            "type": "string",
            "title": "Analysis Timestamp"
          },
          "wallets_found": {
            "type": "integer",
            "title": "Wallets Found"
          },
          "credits_used": {
            "type": "integer",
            "title": "Credits Used"
          },
          "wallets": {
            "items": {
              "$ref": "#/components/schemas/Wallet"
            },
            "type": "array",
            "title": "Wallets"
          }
        },
        "type": "object",
        "required": [
          "id",
          "analysis_timestamp",
          "wallets_found",
          "credits_used",
          "wallets"
        ],
        "title": "AnalysisRun"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "Simple message response"
      },
...
      "RefreshMarketCapsRequest": {
        "properties": {
          "token_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "minItems": 1,
            "title": "Token Ids"
          }
        },
        "type": "object",
        "required": [
          "token_ids"
        ],
        "title": "RefreshMarketCapsRequest"
      },
      "RefreshMarketCapsResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/RefreshMarketCapResult"
            },
            "type": "array",
            "title": "Results"
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens"
          },
          "successful": {
            "type": "integer",
            "title": "Successful"
          },
          "api_credits_used": {
            "type": "integer",
            "title": "Api Credits Used"
          }
        },
        "type": "object",
        "required": [
          "message",
          "results",
          "total_tokens",
          "successful",
          "api_credits_used"
        ],
        "title": "RefreshMarketCapsResponse"
      },
...
      "TokenDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "token_address": {
            "type": "string",
            "title": "Token Address"
          },
          "token_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Name"
          },
          "token_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Symbol"
          },
          "acronym": {
            "type": "string",
            "title": "Acronym"
          },
          "analysis_timestamp": {
            "type": "string",
            "title": "Analysis Timestamp"
          },
          "first_buy_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Buy Timestamp"
          },
          "wallets_found": {
            "type": "integer",
            "title": "Wallets Found"
          },
          "credits_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits Used"
          },
          "last_analysis_credits": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Analysis Credits"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "market_cap_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Usd"
          },
          "market_cap_usd_current": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Usd Current"
          },
          "market_cap_usd_previous": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Usd Previous"
          },
          "market_cap_updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Updated At"
          },
          "market_cap_ath": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Ath"
          },
          "market_cap_ath_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Ath Timestamp"
          },
          "wallets": {
            "items": {
              "$ref": "#/components/schemas/Wallet"
            },
            "type": "array",
            "title": "Wallets"
          },
          "axiom_json": {
            "items": {},
            "type": "array",
            "title": "Axiom Json"
          }
        },
        "type": "object",
        "required": [
          "id",
          "token_address",
          "token_name",
          "token_symbol",
          "acronym",
          "analysis_timestamp",
          "first_buy_timestamp",
          "wallets_found",
          "wallets",
          "axiom_json"
        ],
        "title": "TokenDetail",
        "description": "Token with wallet details and axiom data"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Wallet": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "wallet_address": {
            "type": "string",
            "title": "Wallet Address"
          },
          "first_buy_timestamp": {
            "type": "string",
            "title": "First Buy Timestamp"
          },
          "total_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Usd"
          },
          "transaction_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Count"
          },
          "average_buy_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Buy Usd"
          },
          "wallet_balance_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wallet Balance Usd"
          }
        },
        "type": "object",
        "required": [
          "id",
          "wallet_address",
          "first_buy_timestamp",
          "total_usd",
          "transaction_count",
          "average_buy_usd",
          "wallet_balance_usd"
        ],
        "title": "Wallet"
      },
...
    }
  }
}